Service added and other
Now the client is a service. Also, a new module is being prepared to be added that allows loading unsigned drivers. The code from the driverutil folder was tested on Windows 10, and it is also planned to add an implementation on Windows 11. It is not yet finished, so you will have to install the RTcore64 driver yourself. PRA now only works when run as administrator.